done 4-21

+++ Blog/controllers/blogs.js
@@ -60,6 +60,11 @@ blogsRouter.get('/:id', async (request, response) => {
 })
 
 blogsRouter.delete('/:id', async (request, response, next) => {
+  const blog = await Blog.findById(request.params.id)
+  const decodedToken = jwt.verify(request.token, process.env.SECRET)
+  if(blog.user.toString() !== decodedToken.id.toString()) {
+    return response.status(401).json({error: 'token invalid'})
+  }
   await Blog.findByIdAndRemove(request.params.id)
   response.status(204).end()
 })
